add some basic end to end functional tests

+++ android-sdk/FlagshipSdk/src/test/java/com/flagship/sdk/EdgeEvaluatorTest.kt
@@ -868,7 +868,7 @@ class EdgeEvaluatorTest {
             rolloutPercentage = rolloutPercentage,
             rules = rules,
             type = Type.Boolean,
-            updatedAt = System.currentTimeMillis().toDouble(),
+            updatedAt = System.currentTimeMillis(),
             variants = variants,
         )
 }
